Add Windows service

+++ CpuMonitor/Persistence.fs
@@ -6,11 +6,12 @@ open Newtonsoft.Json
 type FileCreateOptions = Overwrite | Archive
 
 let archive path =
-    let directory = Path.GetDirectoryName path
-    let fnwe = Path.GetFileNameWithoutExtension path
-    let extension = Path.GetExtension path
-    let archived = Path.Combine(directory, Path.ChangeExtension(fnwe + Guid.NewGuid().ToString(), extension))
-    File.Move(path, archived)
+    if File.Exists path then
+        let directory = Path.GetDirectoryName path
+        let fnwe = Path.GetFileNameWithoutExtension path
+        let extension = Path.GetExtension path
+        let archived = Path.Combine(directory, Path.ChangeExtension(fnwe + Guid.NewGuid().ToString(), extension))
+        File.Move(path, archived)
 
 let ensureDirectory path =
     let dir = Path.GetDirectoryName path
